+++ R/app_ui.R
@@ -27,16 +27,26 @@ app_ui <- function(request) {
     shinydashboard::dashboardHeader(title = "Shiny app example"),
     shinydashboard::dashboardSidebar(
       collapsed = FALSE,
-      shinydashboard::sidebarMenu(
-        shinydashboard::menuItem("Data", tabName = "data", icon = icon("upload")),
-        shinydashboard::menuItem("Image", tabName = "image", icon = icon("image")),
-        shinydashboard::menuItem("Plot", tabName = "plot", icon = icon("chart-bar")),
-        shinydashboard::menuItem("Table", tabName = "table",icon = icon("table"))
+      div("Modules", style = "padding: 10px"),
+      shinydashboard::sidebarMenu(id = "module_sidebar",
+                                  shinydashboard::menuItem("Upload & Filter", tabName = "upload", icon = icon("upload")),
+                                  shinydashboard::menuItem("Image", tabName = "image", icon = icon("image")),
+                                  shinydashboard::menuItem("Plot", tabName = "plot", icon = icon("chart-bar")),
+                                  shinydashboard::menuItem("Table", tabName = "table",icon = icon("table"))
+      ),
+      div("Resources", style = "padding: 10px"),
+      shinydashboard::sidebarMenu(id = "resource_sidebar",
+                                  shinydashboard::menuItem("Mastering Shiny", icon = icon("book"),
+                                                           href = "https://mastering-shiny.org/", newtab = TRUE),
+                                  shinydashboard::menuItem("Engineering Production-grade", icon = icon("book"),
+                                                           href = "https://engineering-shiny.org/", newtab = TRUE),
+                                  shinydashboard::menuItem("Code", icon = icon("file-code-o", verify_fa = FALSE),
+                                                           href = "https://github.com/gcfrench/shinystore", newtab = TRUE)
       )
     ),
     shinydashboard::dashboardBody(
       shinydashboard::tabItems(
-        shinydashboard::tabItem(tabName = "data", uiOutput("tab1UI")),
+        shinydashboard::tabItem(tabName = "upload", uiOutput("tab1UI")),
         shinydashboard::tabItem(tabName = "image", uiOutput("tab2UI")),
         shinydashboard::tabItem(tabName = "plot", uiOutput("tab3UI")),
         shinydashboard::tabItem(tabName = "table", uiOutput("tab4UI"))